pyproject.toml
[tool.poetry]
name = "na-alliances-discord-bot"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
discord-py = "^2.4.0"
requests = "^2.32.3"
audioop-lts = "^0.2.2"
aiosqlite = "^0.21.0"
gspread = "^6.2.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
